fix(processing): adapt is_padding to fix potential MemoryError
If an unknown chunk size is larger than available RAM on the system
where unblob is run, the previous is_padding implementation could lead
to MemoryError as it tries to load everything in memory.

Fixed by iterating over the unknown chunk with iterate_file while using
an early return optimization so we return as soon as we have different bytes.