chore(nix): re-organize nix packaging
Instead of adding patches on top of unlob package available on nixpkgs,
we keep our own package.nix version here so that it's easy to add new
third party dependencies that new handlers may require.

It also makes it easier to upstream it since we just have to submit the
package.nix file to nixpkgs.