fix: ensure output directory is created before extraction start
Since we introduced sandboxing with landlock, we sometimes had
permission errors when creating content in the extraction root (see
https://github.com/onekey-sec/unblob/issues/1180).

Instead of extending the sandboxing profile to give extractors the
ability to create content in the extraction root parent, we simply make
sure the extraction root directory exists prior to calling process_file
within the sandbox.

This keeps the sandbox permissions to a minimum while getting rid of
permission errors.

Co-authored-by: Quentin Kaiser <quentin.kaiser@onekey.com>