feat(handler): add support for PE files with extraction of NSIS executable
Add support for PE file by relying on LIEF to parse PE file once matched
on 'MZ' or 'PE' signature.

If the file is a self-extractable NSIS executable
("Nullsoft.NSIS.exehead" present in manifest) we extract it with 7zip.

Note: the DLL files within MSI extraction directory are no longer
extracted since the PE handler takes care of them. This is an
improvement over the RAR false positive being found in the DLL.

Co-authored-by: Quentin Kaiser <quentin.kaiser@onekey.com>